requirements.txt
customtkinter==5.1.3
ipython==8.14.0
matplotlib==3.7.1
networkx==3.1
Pillow==9.5.0
pandas==1.5.3
